+++ timm/models/__init__.py
@@ -21,5 +21,5 @@ from .sknet import *
 from .registry import *
 from .factory import create_model
 from .helpers import load_checkpoint, resume_checkpoint
-from .test_time_pool import TestTimePoolHead, apply_test_time_pool
-from .split_batchnorm import convert_splitbn_model
+from .layers import TestTimePoolHead, apply_test_time_pool
+from .layers import convert_splitbn_model
