+++ timm/models/layers/__init__.py
@@ -1,1 +1,8 @@
+from .conv2d_layers import select_conv2d, MixedConv2d, CondConv2d, ConvBnAct, SelectiveKernelConv
 from .eca import EcaModule, CecaModule
+from .activations import *
+from .adaptive_avgmax_pool import \
+    adaptive_avgmax_pool2d, select_adaptive_pool2d, AdaptiveAvgMaxPool2d, SelectAdaptivePool2d
+from .nn_ops import DropBlock2d, DropPath
+from .test_time_pool import TestTimePoolHead, apply_test_time_pool
+from .split_batchnorm import SplitBatchNorm2d, convert_splitbn_model
