+++ timm/models/layers/adaptive_avgmax_pool.py
