+++ timm/models/pnasnet.py
@@ -14,7 +14,7 @@ import torch.nn.functional as F
 
 from .registry import register_model
 from .helpers import load_pretrained
-from .adaptive_avgmax_pool import SelectAdaptivePool2d
+from .layers import SelectAdaptivePool2d
 
 __all__ = ['PNASNet5Large']
