+++ timm/models/res2net.py
@@ -11,7 +11,7 @@ import torch.nn.functional as F
 from .resnet import ResNet, SEModule
 from .registry import register_model
 from .helpers import load_pretrained
-from .adaptive_avgmax_pool import SelectAdaptivePool2d
+from .layers import SelectAdaptivePool2d
 from timm.data import IMAGENET_DEFAULT_MEAN, IMAGENET_DEFAULT_STD
 
 __all__ = []
